+++ packages/components/organisms/f-header/src/components/NavLink.vue
@@ -7,13 +7,13 @@
             [$style['c-navLink']]: !isPopoverLink,
             [$style['c-navLink--popoverLink']]: isPopoverLink,
             [$style['c-nav-list-link--alt']]: isAltColour,
-            [$style['c-navLink--condensedMid']]: isCondensedOnMid,
+            [$style['c-navLink--condensedMid']]: isCondensed,
             [$style['c-navLink--hoverWithWhiteBackground']]: backgroundTheme === 'white',
             [$style['c-nav-list-link--transparent']]: backgroundTheme === 'transparent'
         }]">
         <div
             v-if="hasIcon"
-            :class="[$style['c-navLink-icon'], { [$style['hide-on-mid']]: isCondensedOnMid }]">
+            :class="[$style['c-navLink-icon'], { [$style['hide-on-mid']]: isCondensed }]">
             <slot name="icon" />
         </div>
         <span
@@ -68,7 +68,7 @@ export default {
             default: false
         },
 
-        isCondensedOnMid: {
+        isCondensed: {
             type: Boolean,
             default: false
         },